analyze import err
